Add occupancy data to DataWrangling notebook
I added code to the DataWrangling notebook to import and resample the occupancy data that had been gathered by our team's Flask web app. In addition, I found a problem with the way the original data had been resampled, so I added an explanation for why it needs to be resampled separately from the other sensor data.